apps/agents/requirements.txt
fastapi>=0.115
uvicorn>=0.34
langgraph>=0.4
langchain-openai>=0.3
langchain-core>=0.3
httpx>=0.28
python-dotenv>=1.0
pydantic>=2.10
